requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
bson
gunicorn
